+++ libnavui-dropin/src/test/java/com/mapbox/navigation/dropin/component/marker/FreeDriveLongPressMapComponentTest.kt
@@ -7,6 +7,7 @@ import com.mapbox.maps.plugin.gestures.OnMapLongClickListener
 import com.mapbox.maps.plugin.gestures.gestures
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
+import com.mapbox.navigation.dropin.NavigationViewContext
 import com.mapbox.navigation.dropin.util.HapticFeedback
 import com.mapbox.navigation.dropin.util.TestStore
 import com.mapbox.navigation.testing.MainCoroutineRule
@@ -23,6 +24,8 @@ import io.mockk.unmockkObject
 import io.mockk.verify
 import io.mockk.verifyOrder
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
@@ -34,6 +37,7 @@ internal class FreeDriveLongPressMapComponentTest {
     @get:Rule
     var coroutineRule = MainCoroutineRule()
 
+    private val enableOnMapLongClick = MutableStateFlow<Boolean>(true)
     private val mockGesturesPlugin: GesturesPlugin = mockk(relaxed = true)
     private val mockMapView: MapView = mockk {
         every { gestures } returns mockGesturesPlugin
@@ -41,6 +45,11 @@ internal class FreeDriveLongPressMapComponentTest {
     private val mockMapboxNavigation: MapboxNavigation = mockk(relaxed = true) {
         every { navigationOptions } returns mockk {
             every { applicationContext } returns mockk(relaxed = true)
+        }
+    }
+    private val navigationViewContext: NavigationViewContext = mockk(relaxed = true) {
+        every { options } returns mockk(relaxed = true) {
+            every { enableMapLongClickIntercept } returns enableOnMapLongClick.asStateFlow()
         }
     }
 
@@ -53,7 +62,7 @@ internal class FreeDriveLongPressMapComponentTest {
         every { HapticFeedback.create(any()) } returns mockk(relaxed = true)
         testStore = spyk(TestStore())
 
-        sut = FreeDriveLongPressMapComponent(testStore, mockMapView)
+        sut = FreeDriveLongPressMapComponent(testStore, mockMapView, navigationViewContext)
     }
 
     @After
@@ -78,7 +87,7 @@ internal class FreeDriveLongPressMapComponentTest {
     }
 
     @Test
-    fun `onMapLongClick should update view model state`() {
+    fun `onMapLongClick should update view model state when long click intercept is enabled`() {
         val slot = slot<OnMapLongClickListener>()
         every { mockGesturesPlugin.addOnMapLongClickListener(capture(slot)) } returns Unit
         sut.onAttached(mockMapboxNavigation)
@@ -87,6 +96,24 @@ internal class FreeDriveLongPressMapComponentTest {
         slot.captured.onMapLongClick(point)
 
         verifyOrder {
+            testStore.dispatch(DestinationAction.SetDestination(Destination(point)))
+            testStore.dispatch(
+                NavigationStateAction.Update(NavigationState.DestinationPreview)
+            )
+        }
+    }
+
+    @Test
+    fun `onMapLongClick should not update state when long click intercept is disabled`() {
+        enableOnMapLongClick.value = false
+        val slot = slot<OnMapLongClickListener>()
+        every { mockGesturesPlugin.addOnMapLongClickListener(capture(slot)) } returns Unit
+        sut.onAttached(mockMapboxNavigation)
+
+        val point = Point.fromLngLat(11.0, 12.0)
+        slot.captured.onMapLongClick(point)
+
+        verify(exactly = 0) {
             testStore.dispatch(DestinationAction.SetDestination(Destination(point)))
             testStore.dispatch(
                 NavigationStateAction.Update(NavigationState.DestinationPreview)